requirements.txt
beautifultable==1.1.0
datasets==3.5.0
huggingface_hub==0.30.2
imbalanced_learn==0.13.0
matplotlib==3.10.5
numpy==2.3.2
pandas==2.3.1
Pillow==11.3.0
Requests==2.32.4
scikit_learn>=1.6,<1.7
scipy==1.16.1
seaborn==0.13.2
timm==1.0.19
torch==2.8.0
tqdm==4.67.1
typing_extensions==4.14.1
#umap-learn
opencv-python-headless==4.11.0.86
huggingface_hub==0.30.2